agent/mcp_agent/integrations/mem0_integration.py
# Placeholder for Mem0 integration

import os
from typing import List, Dict, Any, Optional

# Use try-except for mem0 import in case it causes issues
try:
    from mem0 import MemoryClient
except ImportError as e:
    print(f"Failed to import mem0. Memory features will be unavailable: {e}")
    MemoryClient = None # Define as None if import fails

# Automatically load environment variables if python-dotenv is installed
try:
    from dotenv import load_dotenv
    load_dotenv()
except ImportError:
    print("python-dotenv not found, skipping loading .env file")

class Mem0MemoryManager:
    """Manages interactions with the Mem0 service for semantic memory.

    Handles adding, searching, retrieving, updating, and deleting memories
    associated with specific user IDs.
    """

    def __init__(self, api_key: Optional[str] = None):
        """Initializes the Mem0 Memory Manager.

        Args:
            api_key: The Mem0 API key. If None, attempts to read from
                     the MEM0_API_KEY environment variable.

        Raises:
            ValueError: If Mem0 client could not be imported or if the API key is missing.
        """
        if MemoryClient is None:
            raise ValueError("Mem0ai library not installed or failed to import.")

        self.api_key = api_key or os.getenv("MEM0_API_KEY")
        if not self.api_key:
            raise ValueError("Mem0 API key not provided or found in environment variables.")

        # Initialize the client. Note: Mem0 client itself might not be async,
        # but we wrap its calls in async methods for consistency in our agent architecture.
        try:
            self.client = MemoryClient(api_key=self.api_key)
            print("Mem0 Client initialized successfully.")
        except Exception as e:
            print(f"Failed to initialize Mem0 Client: {e}")
            raise ValueError(f"Failed to initialize Mem0 Client: {e}") from e

    async def add_memory(self, user_id: str, content: Optional[str] = None, messages: Optional[List[Dict[str, str]]] = None, metadata: Optional[dict] = None) -> Any:
        """Adds a memory or a list of messages to Mem0 for a specific user.

        Provide either 'content' for a single memory string or 'messages' for a conversation list.

        Args:
            user_id: The identifier for the user whose memory is being added.
            content: A single string content for the memory.
            messages: A list of message dictionaries (e.g., [{'role': 'user', 'content': '...'}]).
            metadata: Optional metadata to associate with the memory.

        Returns:
            The result from the Mem0 client's add operation.

        Raises:
            ValueError: If neither content nor messages are provided.
            Exception: If the Mem0 API call fails.
        """
        if not content and not messages:
            raise ValueError("Either 'content' or 'messages' must be provided.")

        add_kwargs = {"user_id": user_id}
        if metadata:
            add_kwargs["metadata"] = metadata

        try:
            print(f"Adding memory for user_id: {user_id}")
            if messages:
                result = self.client.add(messages=messages, **add_kwargs)
            else: # content must be provided due to the check above
                result = self.client.add(content=content, **add_kwargs)
            print(f"Mem0 add result: {result}")
            return result
        except Exception as e:
            print(f"Error adding memory to Mem0 for user {user_id}: {e}")
            raise

    async def search_memory(self, user_id: str, query: str, limit: int = 5, filters: Optional[Dict] = None) -> List[Dict[str, Any]]:
        """Searches for relevant memories for a specific user based on a query.

        Args:
            user_id: The identifier for the user whose memory is being searched.
            query: The search query string.
            limit: The maximum number of results to return.
            filters: Optional filters to apply to the search.

        Returns:
            A list of memory dictionaries matching the query.

        Raises:
            Exception: If the Mem0 API call fails.
        """
        try:
            print(f"Searching memory for user_id: {user_id} with query: '{query[:50]}...'")
            search_kwargs = {"user_id": user_id, "limit": limit}
            if filters:
                search_kwargs["filters"] = filters

            results = self.client.search(query=query, **search_kwargs)
            print(f"Mem0 search returned {len(results)} results.")
            return results
        except Exception as e:
            print(f"Error searching Mem0 memory for user {user_id}: {e}")
            raise

    async def get_all_memories(self, user_id: str) -> List[Dict[str, Any]]:
        """Retrieves all memories associated with a specific user.

        Args:
            user_id: The identifier for the user.

        Returns:
            A list of all memory dictionaries for the user.

        Raises:
            Exception: If the Mem0 API call fails.
        """
        try:
            print(f"Getting all memories for user_id: {user_id}")
            memories = self.client.get_all(user_id=user_id)
            print(f"Mem0 get_all returned {len(memories)} memories.")
            return memories
        except Exception as e:
            print(f"Error getting all memories from Mem0 for user {user_id}: {e}")
            raise

    async def get_memory(self, memory_id: str) -> Optional[Dict[str, Any]]:
        """Retrieves a specific memory by its ID.

        Args:
            memory_id: The unique identifier of the memory.

        Returns:
            The memory dictionary if found, otherwise potentially None (depends on client behavior).

        Raises:
            Exception: If the Mem0 API call fails.
        """
        try:
            print(f"Getting memory with id: {memory_id}")
            memory = self.client.get(memory_id=memory_id)
            print(f"Mem0 get returned: {memory is not None}")
            return memory
        except Exception as e:
            print(f"Error getting memory {memory_id} from Mem0: {e}")
            raise

    async def update_memory(self, memory_id: str, data: Dict[str, Any]) -> Any:
        """Updates an existing memory.

        Args:
            memory_id: The ID of the memory to update.
            data: A dictionary containing the fields to update.

        Returns:
            The result from the Mem0 client's update operation.

        Raises:
            Exception: If the Mem0 API call fails.
        """
        try:
            print(f"Updating memory with id: {memory_id}")
            result = self.client.update(memory_id=memory_id, data=data)
            print(f"Mem0 update result: {result}")
            return result
        except Exception as e:
            print(f"Error updating memory {memory_id} in Mem0: {e}")
            raise

    async def delete_memory(self, memory_id: str) -> Any:
        """Deletes a specific memory by its ID.

        Args:
            memory_id: The ID of the memory to delete.

        Returns:
            The result from the Mem0 client's delete operation.

        Raises:
            Exception: If the Mem0 API call fails.
        """
        try:
            print(f"Deleting memory with id: {memory_id}")
            result = self.client.delete(memory_id=memory_id)
            print(f"Mem0 delete result: {result}")
            return result
        except Exception as e:
            print(f"Error deleting memory {memory_id} from Mem0: {e}")
            raise

# Note: The Mem0 Python client itself doesn't appear to be inherently async based on docs.
# These async wrappers execute the synchronous client calls within the async function,
# allowing them to be awaited in an async environment (like LangGraph/FastAPI).
# For true async IO with Mem0, one might need an async HTTP client (like httpx)
# to call the Mem0 REST API directly if the Python SDK doesn't support it natively. 